% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_testing.R
\name{bfCorrectV}
\alias{bfCorrectV}
\title{Perform Bonferroni correction on a vector of p-values}
\usage{
bfCorrectV(pvals, nComp)
}
\arguments{
\item{pvals}{A numeric vector.}

\item{nComp}{Number of comparisons. In most situations, this parameter
should not be changed.}
}
\value{
Bonferroni-corrected p-values
}
\description{
This function performs Bonferroni correction on a vector of p-values.
}
\details{
This function is implemented in order to allow users to perform
Bonferroni correction with fewer comparisons than the number of elements
in the vector, which is normally disallowed by the \code{p.adjust} function
from \code{stats}. A use case is correcting the Seurat markers returned
for each identity class by setting the number of comparisons to the number
of classes.
}
\keyword{internal}
