% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_lists.R
\name{buildMarkerList}
\alias{buildMarkerList}
\title{Find markers for Seurat identity class}
\usage{
buildMarkerList(
  seuratObj,
  idClass = "seurat_clusters",
  invert = FALSE,
  logFCThr = 0,
  minPct = 0,
  minPctRatio = 0,
  ids1 = sort(unique(seuratObj[[]][[idClass]])),
  ids2 = NULL,
  ...
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{idClass}{Identity class.}

\item{invert}{Whether to compute downregulated markers.}

\item{logFCThr}{Fold change threshold for testing.}

\item{minPct}{The minimum fraction of in-cluster cells in which tested
genes need to be expressed.}

\item{minPctRatio}{The minimum ratio of in-cluster cells over out-cluster
cells in which a retained gene must be expressed.}

\item{ids1}{Selected class groups.}

\item{ids2}{Selected class groups used for comparison. Ignored
if \code{invert} is \code{TRUE}.}

\item{...}{Additional arguments passed to FindMarkers.}
}
\value{
A list of marker data frames.
}
\description{
This function finds upregulated or downregulated markers for Seurat object
for a given identity class and performs an additional Bonferroni correction
for multiple testing.
}
