% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_demo.R
\name{buildSeuratMarkerList}
\alias{buildSeuratMarkerList}
\title{Build a Seurat marker list ready to be used by LISTO}
\usage{
buildSeuratMarkerList(seuratObj, col, logFCThr = 1, minPct = 0.2, ...)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{col}{Seurat metadata column used for grouping.}

\item{logFCThr}{Fold change threshold for testing.}

\item{minPct}{The minimum fraction of in-cluster cells in which tested
genes need to be expressed.}

\item{...}{Additional arguments passed to \code{Seurat::FindMarkers}.}
}
\value{
A list consisting of data frames generated with
\code{Seurat::FindMarkers}.
}
\description{
This function builds a Seurat marker list ready to be used by LISTO.
Requires Seurat (not automatically installed with LISTO).
}
\examples{
seuratPath <- system.file('extdata', 'seuratObj.qs2', package='LISTO')
seuratObj <- qs2::qs_read(seuratPath)
a <- buildSeuratMarkerList(seuratObj, 'Cell_Cycle', logFCThr=0.1)

}
