% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_overlaps.R
\name{erDFPathwayOverlap}
\alias{erDFPathwayOverlap}
\title{Compute the overlaps of different pathways for two enrichment results}
\usage{
erDFPathwayOverlap(df, mtMethod = c("by", "bf", "bh"), ...)
}
\arguments{
\item{df}{A data frame joining two \code{enrichResult} objects generated
with \code{joinER}}

\item{mtMethod}{Multiple testing correction method. Choices are
Bonferroni ('bf'), Benjamini-Hochberg('bh'), and the default
Benjamini-Yekutieli ('by').}

\item{...}{Additional arguments passed to the multiple testing correction
function. See \code{?hammers::mtCorrectDF} for information.}
}
\value{
A data frame of pathways with significant gene overlaps.
}
\description{
This function compute the overlaps of different pathways for
two \code{enrichResult} objects.
}
\examples{
m1 <- genesER(c('AURKA', 'PTTG2', 'MKI67', 'RRM2', 'BUB1', 'KIF20C'),
'human')
m2 <- genesER(c('AURKA', 'TOP2A', 'CENPF', 'BIRC5', 'BUB1', 'KIF20C'),
'human')
df <- joinER(m1@result, m2@result)
res <- erDFPathwayOverlap(df)

}
