% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_decomposition.R
\name{factorialPrimePowers}
\alias{factorialPrimePowers}
\title{Generate the prime factor decomposition of n factorial.}
\usage{
factorialPrimePowers(n)
}
\arguments{
\item{n}{A positive integer.}
}
\value{
A vector in which positions represent prime numbers (that is, the
first position corresponds to 2, the second position corresponds to 3,
the third position corresponds to 5, etc.) and values
represent their exponents in the factorial decomposition.
}
\description{
This function generates the prime factor decomposition of n factorial.
}
\examples{
factorialPrimePowers(8)

}
