% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_objects.R
\name{filterItems}
\alias{filterItems}
\title{Filter items based on a provided cutoff}
\usage{
filterItems(obj, numCol = NULL, cutoff = NULL, compFun = `>`)
}
\arguments{
\item{obj}{A data frame with a numeric column, or a character vector.}

\item{numCol}{The name of the numeric column used for data frame ordering.}

\item{cutoff}{Cutoff for assessing item overlaps.}

\item{compFun}{Comparison function.}
}
\description{
This function filters items based on a provided cutoff.
}
\keyword{internal}
