% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_lists.R
\name{findDownMarkers}
\alias{findDownMarkers}
\title{Find downregulated markers for Seurat identity class}
\usage{
findDownMarkers(
  seuratObj,
  idClass = "seurat_clusters",
  logfcThreshold = 0,
  identities = sort(unique(seuratObj[[]][[idClass]]))
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{idClass}{Identity class.}

\item{logfcThreshold}{Fold change threshold for testing.}

\item{identities}{Selected class groups.}
}
\description{
This function finds downregulated markers for Seurat object for a given
identity class and performs an additional Bonferroni correction for
multiple testing.
}
