% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_lists.R
\name{findUpMarkers}
\alias{findUpMarkers}
\title{Find upregulated markers for Seurat identity class}
\usage{
findUpMarkers(
  seuratObj,
  idClass = "seurat_clusters",
  logfcThreshold = 0,
  identities = sort(unique(seuratObj[[idClass]]))
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{idClass}{Identity class.}

\item{logfcThreshold}{Fold change threshold for testing.}

\item{identities}{Selected class groups.}
}
\value{
A list of marker data frames.
}
\description{
This function finds upregulated markers for Seurat object for a given
identity class and performs an additional Bonferroni correction for
multiple testing.
}
