% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_two_universes.R
\name{geneOverlapProbMNk}
\alias{geneOverlapProbMNk}
\title{Compute the probability of intersection of two gene subsets of sets M and N}
\usage{
geneOverlapProbMNk(m, n, a, b)
}
\arguments{
\item{m}{A gene set.}

\item{n}{A gene set.}

\item{a}{A subset of gene set M.}

\item{b}{A subset of gene set N.}
}
\value{
The probability of intersection of the two gene subsets.
}
\description{
This function computes the probability of intersection of two gene subsets
of sets M and N.
}
\details{
A thin wrapper around \code{subsetOverlapProbMNk}.
}
\examples{
geneOverlapProbMNk(LETTERS[seq(19)],
LETTERS[seq(6, 26)],
LETTERS[seq(4, 10)],
LETTERS[seq(7, 15)])

}
