% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoffs.R
\name{generateCutoffs}
\alias{generateCutoffs}
\title{Generate cutoffs for filtering overlaps}
\usage{
generateCutoffs(
  obj1,
  obj2,
  obj3 = NULL,
  numCol = NULL,
  isHighTop = TRUE,
  maxCutoffs = 5000
)
}
\arguments{
\item{obj1}{A data frame with a numeric column, or a character vector.}

\item{obj2}{A data frame with a numeric column, or a character vector.}

\item{obj3}{A data frame with a numeric column, or a character vector.}

\item{numCol}{The name of the numeric column used for data frame ordering.}

\item{isHighTop}{Whether higher values in the numeric column correspond to
top-ranked items.}

\item{maxCutoffs}{Maximum number of cutoffs. If the input data frames
contain more cutoffs than this value, only \code{maxCutoffs} linearly
spaced cutoffs will be selected from the generated cutoff list.}
}
\value{
A numeric vector.
}
\description{
This function generates cutoffs for filtering overlaps
}
\keyword{internal}
