% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoffs.R
\name{getObjectValues}
\alias{getObjectValues}
\title{Extract numeric values from an input object}
\usage{
getObjectValues(obj, numCol = NULL, isHighTop = TRUE)
}
\arguments{
\item{obj}{A data frame with a numeric column, or a character vector.}

\item{numCol}{The name of the numeric column used for data frame ordering.}

\item{isHighTop}{Whether higher values in the numeric column correspond to
top-ranked items.}
}
\description{
This function extracts numeric values from an input object.
}
\keyword{internal}
