% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{joinER}
\alias{joinER}
\title{Join the data frames of two enrichment results}
\usage{
joinER(
  df1,
  df2,
  sharedCols = c("Description", "BgRatio"),
  specificCols = c("p.adjust", "geneID")
)
}
\arguments{
\item{df1}{The \code{result} data frame of an \code{enrichResult} object.}

\item{df2}{The \code{result} data frame of an \code{enrichResult} object.}

\item{sharedCols}{Columns from the \code{enrichResult} objects used for the
join that have the same values for each term shared by the two objects.}

\item{specificCols}{Columns from the \code{enrichResult} objects used for the
join that have potentially different values for terms shared by the
two objects.}
}
\value{
Genes enriched for terms.
}
\description{
This function joins the data frames of two \code{enrichResult} objects.
}
\examples{
m1 <- genesER(c('AURKA', 'PTTG2', 'MKI67', 'RRM2', 'BUB1', 'KIF20C'),
'human')
m2 <- genesER(c('AURKA', 'TOP2A', 'CENPF', 'BIRC5', 'BUB1', 'KIF20C'),
'human')
df <- joinER(m1@result, m2@result)


}
