% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_overlaps.R
\name{markerDFPairOverlap}
\alias{markerDFPairOverlap}
\title{Assess the overlap of two marker data frames.}
\usage{
markerDFPairOverlap(
  markers1,
  markers2,
  genes1,
  genes2 = NULL,
  nCores = 1,
  colStr = "avg_log2FC",
  isHighTop = TRUE,
  extraCutoff = 0,
  maxNCutoffs = 500,
  mtMethod = c("by", "bh"),
  verbose = FALSE
)
}
\arguments{
\item{markers1}{Data frame of markers with at least one numeric column.}

\item{markers2}{Data frame of markers with at least one numeric column.}

\item{genes1}{Genes in the first dataset.}

\item{genes2}{Genes in the second dataset. If \code{NULL} (as default), no
second dataset will be used.}

\item{nCores}{Number of cores. Only used if \code{nDatasets} is 2.}

\item{colStr}{Numeric column.}

\item{isHighTop}{Whether higher values in the numeric column correspond to
top markers.}

\item{extraCutoff}{Cutoff to be placed at one end of the cutoff list.
If \code{isHighTop} is \code{TRUE}, it must be lower than all the cutoffs
present in the marker list, and if \code{isHighTop} is \code{FALSE}, it must
be higher that all cutoffs present in the marker list.}

\item{maxNCutoffs}{Maximum number of cutoffs. If the two input data frames
contain more cutoffs than this value, only \code{maxNCutoffs} linearly
spaced cutoffs will be selected from the original cutoff list.}

\item{mtMethod}{Multiple testing correction method. Choose between
Benjamini-Yekutieli ('by') and Benjamini-Hochberg ('bh').}

\item{verbose}{Whether the output should be verbose.}
}
\value{
A numeric value (hypergeometric p-value).
}
\description{
This function assesses the overlap of two marker data frames.
}
