% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_overlaps.R
\name{markerDFPairPval}
\alias{markerDFPairPval}
\title{Compute the p-value of overlap for two marker data frames}
\usage{
markerDFPairPval(
  markers1,
  markers2,
  colStr,
  cutoffs,
  nDatasets = c("1", "2"),
  nCores = 1,
  mtMethod = c("by", "bh"),
  ...
)
}
\arguments{
\item{markers1}{Data frame of markers with at least one numeric column.}

\item{markers2}{Data frame of markers with at least one numeric column.}

\item{colStr}{Numeric column.}

\item{cutoffs}{Cutoffs for assessing marker overlaps.}

\item{nDatasets}{Number of datasets.}

\item{nCores}{Number of cores. Only used if \code{nDatasets} is 2.}

\item{mtMethod}{Multiple testing correction method. Choose between
Benjamini-Yekutieli ('by') and Benjamini-Hochberg ('bh').}

\item{...}{Additional parameters passed to \code{pvalOverlap} or
\code{pvalOverlapMN}.}
}
\value{
A p-value.
}
\description{
This function computes the p-value of overlap for two marker data frames.
}
\keyword{internal}
