% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_overlaps.R
\name{markerListPhyper}
\alias{markerListPhyper}
\title{Assess the overlap of two lists of marker data frames.}
\usage{
markerListPhyper(
  markerList1,
  markerList2,
  nGenes,
  colStr = "avg_log2FC",
  isHighTop = TRUE,
  extraCutoff = 0,
  pvalThr = 0.05,
  verbose = TRUE
)
}
\arguments{
\item{markerList1}{List of data frames.}

\item{markerList2}{List of data frames.}

\item{nGenes}{Number of genes in the dataset.}

\item{colStr}{Numeric column.}

\item{isHighTop}{Whether higher values in the numeric column correspond to
top markers.}

\item{extraCutoff}{Cutoff to be placed at one end of the cutoff list.
If \code{isHighTop} is \code{TRUE}, it must be lower than all the cutoffs
present in the marker list, and if \code{isHighTop} is \code{FALSE}, it must
be higher that all cutoffs present in the marker list.}

\item{pvalThr}{p-value threshold.}

\item{verbose}{Whether the output should be verbose.}
}
\value{
A data frame.
}
\description{
This function assesses the overlap of two lists of marker data frames.
}
