% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{markerSetsPhyper}
\alias{markerSetsPhyper}
\title{Assess the overlap of two marker data frames.}
\usage{
markerSetsPhyper(
  markers1,
  markers2,
  nGenes,
  colStr = "avg_log2FC",
  isHighTop = TRUE,
  extraCutoff = 0
)
}
\arguments{
\item{markers1}{Data frame.}

\item{markers2}{Data frame.}

\item{nGenes}{Number of genes in the dataset.}

\item{colStr}{Numeric column.}

\item{isHighTop}{Whether higher values in the numeric column correspond to
top markers.}

\item{extraCutoff}{Cutoff to be placed at one end of the cutoff list.
If \code{isHighTop} is \code{TRUE}, it must be lower than all the cutoffs
present in the marker list, and if \code{isHighTop} is \code{FALSE}, it must
be higher that all cutoffs present in the marker list.}
}
\value{
A numeric value (hypergeometric p-value).
}
\description{
This function assesses the overlap of two marker data frames.
}
