% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_testing.R
\name{mtCorrectDF}
\alias{mtCorrectDF}
\title{Perform multiple testing correction on a data frame}
\usage{
mtCorrectDF(
  df,
  mtMethod = c("BY", "holm", "hochberg", "hommel", "bonferroni", "BH", "fdr", "none"),
  colStr = "pval",
  newColStr = "pvalAdj",
  doOrder = TRUE,
  doFilter = TRUE,
  pvalThr = 0.05,
  ...
)
}
\arguments{
\item{df}{A data frame with a p-values columnn.}

\item{mtMethod}{Multiple testing correction method. Choices are
'BY' (default) 'holm', hochberg', hommel', 'bonferroni', 'BH',  'fdr' and
'none'.}

\item{colStr}{Name of the column of p-values.}

\item{newColStr}{Name of the column of adjusted p-values that will be
created.}

\item{doOrder}{Whether to increasingly order the data frame based on the
adjusted p-values.}

\item{doFilter}{Whether to filter the data frame based on the adjusted
p-values.}

\item{pvalThr}{p-value threshold used for filtering. Ignored
if \code{doFilter} is \code{FALSE}.}

\item{...}{Additional arguments passed to the multiple testing correction
method.}
}
\value{
A data frame in which the p-value column has been corrected for
multiple testing.
}
\description{
This function orders a data frame based on a column of p-values, performs
multiple testing correction on the column, and filters the data-frame
based on the adjusted p-values.
}
\examples{
df <- data.frame(elem = c('A', 'B', 'C', 'D', 'E'),
pval = c(0.032, 0.001, 0.0045, 0.051, 0.048))
mtCorrectDF(df)

}
