% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_testing.R
\name{mtCorrectHelper}
\alias{mtCorrectHelper}
\title{Helper function for multiple comparison testing}
\usage{
mtCorrectHelper(pvals, mtMethod, nComp)
}
\arguments{
\item{pvals}{A numeric vector.}

\item{mtMethod}{Multiple testing correction method. Choices are
'BY' (default) 'holm', hochberg', hommel', 'bonferroni', 'BH',  'fdr' and
'none'.}

\item{nComp}{Number of comparisons. In most situations, this parameter
should not be changed.}
}
\value{
Adjusted p-values.
}
\description{
This function is a helper for multiple comparison testing.
}
\keyword{internal}
