% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_testing.R
\name{mtCorrectV}
\alias{mtCorrectV}
\title{Perform multiple testing correction on a vector of p-values}
\usage{
mtCorrectV(
  pvals,
  mtMethod = c("BY", "holm", "hochberg", "hommel", "bonferroni", "BH", "fdr", "none"),
  mtStat = c("identity", "median", "mean", "max", "min"),
  nComp = length(pvals)
)
}
\arguments{
\item{pvals}{A numeric vector.}

\item{mtMethod}{Multiple testing correction method. Choices are
'BY' (default) 'holm', hochberg', hommel', 'bonferroni', 'BH',  'fdr' and
'none'.}

\item{mtStat}{A statistics to be optionally computed. Choices are 'identity'
(no statistics will be computed and the adjusted p-values will be returned
as such), 'median', 'mean', 'max' and 'min'.}

\item{nComp}{Number of comparisons. In most situations, this parameter
should not be changed.}
}
\value{
If \code{mtStat} is 'identity' (as default), a numeric vector of
p-values corrected for multiple testing. Otherwise, a statistic based on
these corrected p-values defined by \code{mtStat}.
}
\description{
This function performs multiple testing correction on a vector of p-values.
}
\examples{
pvals <- c(0.032, 0.001, 0.0045, 0.051, 0.048)
mtCorrectV(pvals)

}
