% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_two_universes.R
\name{powerProduct}
\alias{powerProduct}
\title{Compute the prime factor decomposition of the binomial coefficient}
\usage{
powerProduct(primes, exponents)
}
\arguments{
\item{n}{Total number of elements.}

\item{k}{Number of selected elements.}
}
\value{
A vector in which positions represent prime numbers and values
represent their exponents in the factorial decomposition.
}
\description{
This function computes the prime factor decomposition of the
binomial coefficient.
}
\examples{
powerProduct(c(2, 3, 5), c(4, 2, 6))

}
