% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_counts.R
\name{probCounts2MN}
\alias{probCounts2MN}
\title{Compute the probability that two subsets of sets M and N intersect in k
points}
\usage{
probCounts2MN(intMN, intAN, intBM, k)
}
\arguments{
\item{intMN}{Number of elements in the intersection of sets M and N.}

\item{intAN}{Number of elements in the intersection of sets A (subset of M)
and N.}

\item{intBM}{Number of elements in the intersection of sets B (subset of N)
and M.}

\item{k}{Number of elements in the intersection of sets A and B.}
}
\value{
A numeric value in [0, 1] representing the probability that two
subsets of sets M and N intersect in k points.
}
\description{
This function computes the probability that two subsets of sets M and N
intersect in k points. Intersection sizes (M with N, A with N and B with M)
must be provided.
}
\examples{
probCounts2MN(8, 6, 4, 2)

}
