% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_counts.R
\name{pvalCounts3N}
\alias{pvalCounts3N}
\title{Compute the probability that three subsets of a set intersect in at least k points}
\usage{
pvalCounts3N(lenA, lenB, lenC, n, k)
}
\arguments{
\item{lenA}{Size of the first subset.}

\item{lenB}{Size of the second subset.}

\item{lenC}{Size of the third subset.}

\item{n}{Size of the set comprising the subsets.}

\item{k}{Size of the intersection.}
}
\value{
A numeric value in [0, 1] representing the probability that
three subsets of a set intersect in at least k points.
}
\description{
This function computes the probability that three subsets of a set intersect
in at least k points.
}
\examples{
pvalCounts3N (300, 200, 250, 400, 180)

}
