% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_counts.R
\name{pvalCountsMN}
\alias{pvalCountsMN}
\title{Compute the probability that two subsets of sets M and N intersect in
at least k points}
\usage{
pvalCountsMN(intMN, intAN, intBM, k)
}
\arguments{
\item{intMN}{Number of elements in the intersection of sets M and N.}

\item{intAN}{Number of elements in the intersection of sets A (subset of M)
and N.}

\item{intBM}{Number of elements in the intersection of sets B (subset of N)
and M.}

\item{k}{Number of elements in the intersection of sets A and B.}
}
\value{
The probability that two subsets of sets M and N intersect in
at least k points.
}
\description{
This function computes the probability that two subsets A and B of sets
M and N intersect in at least k points. In addition to k, the intersection
sizes of M and M, A and N, and B and M must pe provided as input.
}
\examples{
pvalCountsMN (300, 23, 24, 6)

}
