% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_lists.R
\name{pvalLists}
\alias{pvalLists}
\title{Assess the overlap of two lists of marker data frames.}
\usage{
pvalLists(
  list1,
  list2,
  col,
  allItems1,
  allItems2 = NULL,
  nCores = 1,
  isHighTop = TRUE,
  extraCutoff = 0,
  maxNCutoffs = 500,
  mtMethod = c("BY", "holm", "hochberg", "hommel", "bonferroni", "BH", "fdr", "none"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{list1}{List of data frames.}

\item{list2}{List of data frames.}

\item{allItems1}{Either items in the first dataset or their number. If a
second dataset is provided (that is, \code{allItems} is not \code{NULL}),
the items in the first dataset rather than their number must be provided
here.}

\item{allItems2}{All items in the second dataset. If \code{NULL}
(as default), no second dataset will be used.}

\item{nCores}{Number of cores. Only used if \code{nDatasets} is 2.}

\item{isHighTop}{Whether higher values in the numeric column correspond to
top-ranked items.}

\item{mtMethod}{Multiple testing correction method.}

\item{...}{Additional arguments passed to \code{mtCorrectDF}.}
}
\value{
A data frame.
}
\description{
This function assesses the overlap of two lists of marker data frames.
}
