% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_objects.R
\name{pvalObjects}
\alias{pvalObjects}
\title{Assess the overlap of two or three objects}
\usage{
pvalObjects(
  obj1,
  obj2,
  obj3 = NULL,
  universe1,
  universe2 = NULL,
  numCol = NULL,
  isHighTop = TRUE,
  maxCutoffs = 5000,
  mtMethod = c("BY", "holm", "hochberg", "hommel", "bonferroni", "BH", "fdr", "none"),
  nCores = 1,
  type = c("2N", "2MN", "3N")
)
}
\arguments{
\item{obj1}{A data frame with a numeric column, or a character vector.}

\item{obj2}{A data frame with a numeric column, or a character vector.}

\item{obj3}{A data frame with a numeric column, or a character vector.}

\item{universe1}{The set from which the items stored
in \code{obj1} are selected.}

\item{universe2}{The set from which the items stored
in \code{obj2} are selected.}

\item{numCol}{The name of the numeric column used for data frame ordering.}

\item{isHighTop}{Whether higher values in the numeric column correspond to
top-ranked items.}

\item{maxCutoffs}{Maximum number of cutoffs. If the input data frames
contain more cutoffs than this value, only \code{maxCutoffs} linearly
spaced cutoffs will be selected from the generated cutoff list.}

\item{mtMethod}{Multiple testing correction method.}

\item{nCores}{Number of cores. If performing an overlap assessment between
sets belonging to the same universe, it is recommended not to use
parallelization (that is, leave this parameter as 1).}

\item{type}{Type of overlap assessment. Choose between: two sets belonging
to the same universe ('2N'), two sets belonging to different universes
('2MN'), three sets belonging to the same universe ('3MN').}
}
\value{
A numeric value in [0, 1] representing the p-value of the
overlap of the two objects.
}
\description{
This function assesses the overlap of two or three objects
(character vectors, or data frames having a numeric column).
}
\examples{
pvalObjects(LETTERS[seq(2, 7)], LETTERS[seq(3, 19)], universe1=LETTERS)

}
