% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_objects.R
\name{pvalObjectsCore}
\alias{pvalObjectsCore}
\title{Compute the p-value of overlap for two or three objects}
\usage{
pvalObjectsCore(
  obj1,
  obj2,
  obj3 = NULL,
  universe1,
  universe2 = NULL,
  numCol = NULL,
  cutoff = NULL,
  compFun = `>`,
  type = c("2N", "2MN", "3N")
)
}
\arguments{
\item{obj1}{A data frame with a numeric column, or a character vector.}

\item{obj2}{A data frame with a numeric column, or a character vector.}

\item{obj3}{A data frame with a numeric column, or a character vector.}

\item{universe1}{The set from which the items stored
in \code{obj1} are selected.}

\item{universe2}{The set from which the items stored
in \code{obj2} are selected.}

\item{numCol}{The name of the numeric column used for data frame ordering.}
}
\value{
A p-value.
}
\description{
This function computes the p-value of overlap for two or three objects.
}
\keyword{internal}
