% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_functions.R
\name{pvalOverlap}
\alias{pvalOverlap}
\title{Calculate the hypergeometric p-value of enrichment for two sets}
\usage{
pvalOverlap(a, b, n, lowerTail = FALSE)
}
\arguments{
\item{a}{Character vector.}

\item{b}{Chracter vector.}

\item{n}{Number of elements of set from which \code{a} and \code{b} are
selected.}

\item{lowerTail}{Whether to calculate underenrichment (\code{TRUE}) or
overenrichment (\code{FALSE}).}
}
\description{
This function calculates the hypergeometric p-value of enrichment for
two sets.
}
\details{
This function calculates the hypergeometric p-value of enrichment
for two sets.
}
