% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_functions.R
\name{pvalOverlapMN}
\alias{pvalOverlapMN}
\title{Compute the p-value of intersection of two gene subsets of sets M and N}
\usage{
pvalOverlapMN(a, b, m, n)
}
\arguments{
\item{a}{A subset of gene set M.}

\item{b}{A subset of gene set N.}

\item{m}{A gene set.}

\item{n}{A gene set.}
}
\value{
The probability of intersection of the two gene subsets.
}
\description{
This function computes the p-value of intersection of two gene subsets
of sets M and N.
}
\details{
A thin wrapper around \code{pvalOverlapMNk}.
}
\examples{
pvalOverlapMN(LETTERS[seq(4, 10)],
LETTERS[seq(7, 15)],
LETTERS[seq(19)],
LETTERS[seq(6, 26)])

}
