% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_sets.R
\name{pvalSets2N}
\alias{pvalSets2N}
\title{Calculate the p-value of intersection for two sets}
\usage{
pvalSets2N(a, b, n)
}
\arguments{
\item{a}{A character vector.}

\item{b}{A character vector.}

\item{n}{Set from which \code{a} and \code{b} are
selected.}
}
\value{
A numeric value in [0, 1] representing the p-value of intersection
for two sets.
}
\description{
This function calculates the p-value of intersection for two sets.
}
\details{
A thin wrapper around \code{stats::phyper}.
}
\examples{
pvalSets2N(LETTERS[seq(4, 10)], LETTERS[seq(7, 15)], LETTERS)

}
