% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_sets.R
\name{pvalSets3N}
\alias{pvalSets3N}
\title{Compute the p-value of intersection of three subsets}
\usage{
pvalSets3N(a, b, c, n)
}
\arguments{
\item{a}{A character vector.}

\item{b}{A character vector.}

\item{c}{A character vector.}

\item{n}{Set from which \code{a}, \code{b} and \code{c} are
selected.}
}
\value{
A numeric value in [0, 1] representing the p-value of intersection
of three subsets.
}
\description{
This function computes the p-value of intersection of three subsets.
}
\details{
A thin wrapper around \code{pvalCounts3N}.
}
\examples{
pvalSets3N(LETTERS[seq(4, 10)],
LETTERS[seq(7, 15)],
LETTERS[seq(19)],
LETTERS)

}
