% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_subsets.R
\name{pvalSubsetsMN}
\alias{pvalSubsetsMN}
\title{Compute the p-value of intersection of two subsets of sets M and N}
\usage{
pvalSubsetsMN(a, b, m, n)
}
\arguments{
\item{a}{A character vector.}

\item{b}{A character vector.}

\item{m}{Set from which \code{a} is selected.}

\item{n}{Set from which \code{b} is selected.}
}
\value{
The probability of intersection of the two subsets.
}
\description{
This function computes the p-value of intersection of two subsets
of sets M and N.
}
\details{
A thin wrapper around \code{pvalOverlapMNk}.
}
\examples{
pvalSubsetsMN(LETTERS[seq(4, 10)],
LETTERS[seq(7, 15)],
LETTERS[seq(19)],
LETTERS[seq(6, 26)])

}
