% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_aux.R
\name{sharedMarkers}
\alias{sharedMarkers}
\title{Join two marker data frames}
\usage{
sharedMarkers(markerSet1, markerSet2, joinCol = "avg_log2FC")
}
\arguments{
\item{markerSet1}{A marker data frame.}

\item{markerSet2}{A marker data frame.}

\item{joinCol}{Join column.}
}
\value{
A data frame with the shared markers as rows and the join column
from each of the two marker data frames as columns.
}
\description{
This function joins two marker data frames.
}
