% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{sharedMarkersPlot}
\alias{sharedMarkersPlot}
\title{Plots marker shared between two marker data frames}
\usage{
sharedMarkersPlot(
  markerObj,
  title = NULL,
  markerNames = NULL,
  joinCol = "avg_log2FC",
  thresh1 = NULL,
  thresh2 = NULL,
  isNeg1 = FALSE,
  isNeg2 = FALSE,
  nameSuffix1 = NULL,
  nameSuffix2 = NULL,
  ...
)
}
\arguments{
\item{markerObj}{A list of two named marker data frames.}

\item{title}{Plot title.}

\item{markerNames}{Names of markers to be displayed on the plot.
If \code{NULL}, the markers will be chosen based on the thresholds.}

\item{joinCol}{Column based on which the marker data frames will be joined.}

\item{thresh1}{Join column threshold for the first marker data frame.}

\item{thresh2}{Join column threshold for the second marker data frame.}

\item{isNeg1}{Whether the first marker set represents negative markers.}

\item{isNeg2}{Whether the second marker set represents negative markers.}

\item{nameSuffix1}{Suffix appended to \code{name1} on axes and legend.}

\item{nameSuffix2}{Suffix appended to \code{name2} on axes and legend.}

\item{...}{Additional arguments passed to \code{henna::hullPlot}.}
}
\value{
A ggplot object.
}
\description{
This function plots markers shared between two marker data frames.
}
\details{
A wrapper around \code{henna::hullPlot}
}
