% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_two_universes.R
\name{subsetOverlapProbMNk}
\alias{subsetOverlapProbMNk}
\title{Compute the probability that two subsets of sets M and N intersect in k
points}
\usage{
subsetOverlapProbMNk(intMN, intAN, intBM, k)
}
\arguments{
\item{intMN}{Number of elements in the intersection of sets M and N.}

\item{intAN}{Number of elements in the intersection of sets A (subset of M)
and N.}

\item{intBM}{Number of elements in the intersection of sets B (subset of N)
and M.}

\item{k}{Number of elements in the intersection of sets A and B.}
}
\value{
The probability that two subsets of sets M and N intersect in k
points.
}
\description{
This function computes the probability that two subsets of sets M and N
intersect in k points.
}
\examples{
subsetOverlapProbMNk(8, 6, 4, 2)

}
