% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{vChoose}
\alias{vChoose}
\title{Compute the prime factor decomposition of the binomial coefficient}
\usage{
vChoose(n, k)
}
\arguments{
\item{n}{Total number of elements.}

\item{k}{Number of selected elements.}
}
\value{
A vector in which positions represent prime numbers (that is, the
first position corresponds to 2, the second position corresponds to 3,
the third position corresponds to 5, etc.) and values
represent their exponents in the factorial decomposition.
}
\description{
This function computes the prime factor decomposition of the
binomial coefficient.
}
\examples{
vChoose(8, 4)

}
