% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{vNumeratorMN}
\alias{vNumeratorMN}
\title{Compute the prime representation of the numerator of the fraction
representing the probability that two subsets of sets M and N intersect
in k points}
\usage{
vNumeratorMN(intMN, intAN, intBM, k)
}
\arguments{
\item{intMN}{Number of elements in the intersection of sets M and N.}

\item{intAN}{Number of elements in the intersection of sets A (subset of M)
and N.}

\item{intBM}{Number of elements in the intersection of sets B (subset of N)
and M.}

\item{k}{Number of elements in the intersection of sets A and B.}
}
\value{
A vector containing the prime representation of the fraction
representing the probability that two subsets of sets M and N intersect in k
points. Positions represent prime numbers in order (2, 3, 5...), and values
represent their exponents in the prime decomposition.
}
\description{
This function computes the numerator of the fraction representing the
probability that two subsets of sets M and N intersect in k points
}
\keyword{internal}
