% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{vSum}
\alias{vSum}
\title{Add numeric vectors of different lenghts}
\usage{
vSum(...)
}
\arguments{
\item{...}{Numeric vectors.}
}
\value{
A numeric vector.
}
\description{
This function adds numeric vectors of different lengths by filling shorter
vectors with zeroes.
}
\examples{
vSum(c(1, 4), c(2, 8, 6), c(1, 7), c(10, 4, 6, 7))

}
