% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_universes.R
\name{factorialPrimePowers}
\alias{factorialPrimePowers}
\title{Generate the prime factor decomposition of n factorial.}
\usage{
factorialPrimePowers(n)
}
\arguments{
\item{n}{A positive integer.}
}
\value{
A vector in which positions represent prime numbers and values
represent their exponents in the factorial decomposition.
}
\description{
This function generates the prime factor decomposition of n!
}
\examples{
factorialPrimePowers(8)

}
