% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generateCutoffs}
\alias{generateCutoffs}
\title{Generate cutoffs for assessing marker overlaps}
\usage{
generateCutoffs(
  markers1,
  markers2,
  colStr = "avg2_logFC",
  isHighTop = TRUE,
  extraCutoff = 0
)
}
\arguments{
\item{markers1}{Data frame.}

\item{markers2}{Data frame.}

\item{colStr}{Numeric column.}

\item{isHighTop}{Whether higher values in the numeric column correspond to
top markers.}

\item{extraCutoff}{Cutoff to be placed at one end of the cutoff list.
If \code{isHighTop} is \code{TRUE}, it must be lower than all the cutoffs
present in the marker list, and if \code{isHighTop} is \code{FALSE}, it must
be higher that all cutoffs present in the marker list.}
}
\value{
A numeric vector.
}
\description{
This function generates cutoffs for assessing marker overlaps.
}
\keyword{internal}
