% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_overlaps.R
\name{generateCutoffs}
\alias{generateCutoffs}
\title{Generate cutoffs for assessing marker overlaps}
\usage{
generateCutoffs(
  markers1,
  markers2,
  colStr = "avg2_logFC",
  isHighTop = TRUE,
  extraCutoff = 0,
  maxNCutoffs = 10000
)
}
\arguments{
\item{markers1}{Data frame.}

\item{markers2}{Data frame.}

\item{colStr}{Numeric column.}

\item{isHighTop}{Whether higher values in the numeric column correspond to
top markers.}

\item{extraCutoff}{Cutoff to be placed at one end of the cutoff list.
If \code{isHighTop} is \code{TRUE}, it must be lower than all the cutoffs
present in the marker list, and if \code{isHighTop} is \code{FALSE}, it must
be higher that all cutoffs present in the marker list.}

\item{maxNCutoffs}{Maximum number of cutoffs. If the two input data frames
contain more cutoffs than this value, only \code{maxNCutoffs} linearly
spaced cutoffs will be selected from the original cutoff list.}
}
\value{
A numeric vector.
}
\description{
This function generates cutoffs for assessing marker overlaps.
}
\keyword{internal}
