% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_counts.R
\name{probCounts3N}
\alias{probCounts3N}
\title{Compute the probability that three subsets of a set intersect in k
points}
\usage{
probCounts3N(a, b, c, n, k)
}
\arguments{
\item{a}{Size of the first subset.}

\item{b}{Size of the second subset.}

\item{c}{Size of the third subset.}

\item{n}{Size of the set.}

\item{k}{Size of the intersection.}
}
\value{
The probability that three subsets of a set intersect in k points.
}
\description{
This function compute the probability that three subsets of a set intersect
in k points
}
\examples{
probCounts3N(8, 6, 10, 20, 3)

}
