% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_objects.R
\name{pvalObjects}
\alias{pvalObjects}
\title{Assess the overlap of two marker data frames.}
\usage{
pvalObjects(
  obj1,
  obj2,
  obj3 = NULL,
  universe1,
  universe2 = NULL,
  numCol = NULL,
  isHighTop = TRUE,
  maxCutoffs = 500,
  mtMethod = c("BY", "holm", "hochberg", "hommel", "bonferroni", "BH", "fdr", "none"),
  nCores = 1,
  type = c("2N", "2MN", "3N")
)
}
\arguments{
\item{isHighTop}{Whether higher values in the numeric column correspond to
top-ranked items.}

\item{maxCutoffs}{Maximum number of cutoffs. If the input data frames
contain more cutoffs than this value, only \code{maxCutoffs} linearly
spaced cutoffs will be selected from the original cutoff list.}

\item{mtMethod}{Multiple testing correction method.}

\item{nCores}{Number of cores. Only used if \code{nDatasets} is 2.}

\item{allItems1}{Either items in the first dataset or their number. If a
second dataset is provided (that is, \code{allItems} is not \code{NULL}),
the items in the first dataset rather than their number must be provided
here.}

\item{allItems2}{All items in the second dataset. If \code{NULL}
(as default), no second dataset will be used.}
}
\value{
A numeric value (p-value).
}
\description{
This function assesses the overlap of two marker data frames.
}
