% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_objects.R
\name{pvalObjectsCore}
\alias{pvalObjectsCore}
\title{Compute the p-value of overlap for two marker data frames}
\usage{
pvalObjectsCore(
  obj1,
  obj2,
  obj3 = NULL,
  universe1,
  universe2 = NULL,
  numCol = NULL,
  cutoff = NULL,
  compFun = `>`,
  type = c("2N", "2MN", "3N")
)
}
\arguments{
\item{cutoffs}{Cutoffs for assessing item overlaps.}

\item{nDatasets}{Number of datasets. Choose between 1 (the two objects are
selected from the same dataset) and 2 (the two objects are selected
from different datasets).}

\item{nCores}{Number of cores. Only used if \code{nDatasets} is 2.}

\item{mtMethod}{Multiple testing correction method.}

\item{...}{Additional parameters passed to \code{pvalOverlap} or
\code{pvalOverlapMN}.}
}
\value{
A p-value.
}
\description{
This function computes the p-value of overlap for two marker data frames.
}
\keyword{internal}
