% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_objects.R
\name{pvalObjectsCore}
\alias{pvalObjectsCore}
\title{Compute the p-value of overlap for two marker data frames}
\usage{
pvalObjectsCore(
  obj1,
  obj2,
  col,
  cutoffs,
  nDatasets = 1,
  nCores = 1,
  mtMethod = c("BY", "holm", "hochberg", "hommel", "bonferroni", "BH", "fdr", "none"),
  ...
)
}
\arguments{
\item{obj1}{A data frame with at least one numeric column.}

\item{obj2}{A data frame with at least one numeric column.}

\item{col}{Numeric column used for ranking items.}

\item{cutoffs}{Cutoffs for assessing item overlaps.}

\item{nDatasets}{Number of datasets. Choose between 1 (the two objects are
selected from the same dataset) and 2 (the two objects are selected
from different datasets).}

\item{nCores}{Number of cores. Only used if \code{nDatasets} is 2.}

\item{mtMethod}{Multiple testing correction method.}

\item{...}{Additional parameters passed to \code{pvalOverlap} or
\code{pvalOverlapMN}.}
}
\value{
A p-value.
}
\description{
This function computes the p-value of overlap for two marker data frames.
}
\keyword{internal}
