% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_sets.R
\name{pvalSets2N}
\alias{pvalSets2N}
\title{Calculate the hypergeometric p-value of enrichment for two sets}
\usage{
pvalSets2N(a, b, n)
}
\arguments{
\item{a}{A character vector.}

\item{b}{A character vector.}

\item{n}{Set from which \code{a} and \code{b} are
selected.}
}
\description{
This function calculates the hypergeometric p-value of enrichment for
two sets.
}
\details{
This function calculates the hypergeometric p-value of enrichment
for two sets.
}
