% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_listo.R
\name{runLISTO}
\alias{runLISTO}
\title{Assess the overlap of two or three lists of objects.}
\usage{
runLISTO(
  list1,
  list2,
  list3 = NULL,
  universe1,
  universe2 = NULL,
  numCol = NULL,
  isHighTop = TRUE,
  maxCutoffs = 500,
  mtMethod = c("BY", "holm", "hochberg", "hommel", "bonferroni", "BH", "fdr", "none"),
  nCores = 1,
  ...
)
}
\arguments{
\item{list1}{A list containing character vectors, or data frames having
a numeric column.}

\item{list2}{A list containing character vectors, or data frames having a
numeric column.}

\item{list3}{A list containing character vectors, or data frames having a
numeric column.}

\item{universe1}{The set from which the items corresponding to the elements
in \code{list1} are selected.}

\item{universe2}{The set from which the items corresponding to the elements
in \code{list2} are selected.}

\item{numCol}{The name of the numeric column used for data frame ordering.}

\item{isHighTop}{Whether higher values in the numeric column correspond to
top-ranked items.}

\item{maxCutoffs}{Maximum number of cutoffs. If the input data frames
contain more cutoffs than this value, only \code{maxCutoffs} linearly
spaced cutoffs will be selected from the generated cutoff list.}

\item{mtMethod}{Multiple testing correction method.}

\item{nCores}{Number of cores. If performing an overlap assessment between
sets belonging to the same universe, it is recommended not to use
parallelization (that is, leave this parameter as 1).}

\item{...}{Additional arguments passed to \code{mtCorrectDF}.}
}
\value{
A data frame.
}
\description{
This function assesses the overlap of two or three lists of objects
(character vectors, or data frames having at least one numeric column).
}
