% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{sharedMarkersPlot}
\alias{sharedMarkersPlot}
\title{Find markers for Seurat identity class}
\usage{
sharedMarkersPlot(
  markerList1,
  markerList2,
  name1,
  name2,
  joinColumn = "avg_log2FC",
  thresh1 = 1.5,
  thresh2 = 1.5,
  title = paste0("Shared markers - ", name1, " and ", name2),
  ...
)
}
\arguments{
\item{markerList1}{List of data frames.}

\item{markerList2}{List of data frames.}

\item{name1}{Name of first marker data frame.}

\item{name2}{Name of second marker data frame.}

\item{thresh1}{Join column threshold for the first marker data frame.}

\item{thresh2}{Join column threshold for the second marker data frame.}

\item{title}{Plot title}

\item{...}{Additional arguments passed to \code{henna::hullPlot}.}
}
\value{
A list of marker data frames.
}
\description{
This function finds upregulated or downregulate markers for Seurat object
for a given identity class and performs an additional Bonferroni correction for
multiple testing.
}
