% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{sharedMarkersPlot}
\alias{sharedMarkersPlot}
\title{Find markers for Seurat identity class}
\usage{
sharedMarkersPlot(
  markerList1,
  markerList2,
  name1,
  name2,
  markers = NULL,
  joinColumn = "avg_log2FC",
  thresh1 = 1.5,
  thresh2 = 1.5,
  title = "Shared markers",
  markersHullColor = "purple",
  ...
)
}
\arguments{
\item{markerList1}{List of data frames.}

\item{markerList2}{List of data frames.}

\item{name1}{Name of first marker data frame.}

\item{name2}{Name of second marker data frame.}

\item{markers}{Names of markers to be displayed on the plot.
If \code{NULL}, the markers will be chosen based on the thresholds.}

\item{thresh1}{Join column threshold for the first marker data frame.}

\item{thresh2}{Join column threshold for the second marker data frame.}

\item{title}{Plot title.}

\item{markersHullColor}{Color of the hull determined by \code{markers}.
Ignored if \code{markers} is \code{NULL}.}

\item{...}{Additional arguments passed to \code{henna::hullPlot}.}
}
\value{
A list of marker data frames.
}
\description{
This function finds upregulated or downregulate markers for Seurat object
for a given identity class and performs an additional Bonferroni correction for
multiple testing.
}
