% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{sharedMarkersPlot}
\alias{sharedMarkersPlot}
\title{Find markers for Seurat identity class}
\usage{
sharedMarkersPlot(
  markerList1,
  name1,
  name2,
  markerList2 = NULL,
  markerNames = NULL,
  joinColumn = "avg_log2FC",
  thresh1 = NULL,
  thresh2 = NULL,
  title = "Shared markers",
  markersHullColor = "purple",
  isNeg1 = FALSE,
  isNeg2 = FALSE,
  nameSuffix1 = NULL,
  nameSuffix2 = NULL,
  ...
)
}
\arguments{
\item{markerList1}{List of marker data frames.}

\item{name1}{Name of first marker data frame.}

\item{name2}{Name of second marker data frame.}

\item{markerList2}{List of marker data frames.}

\item{markerNames}{Names of markers to be displayed on the plot.
If \code{NULL}, the markers will be chosen based on the thresholds.}

\item{joinColumn}{Column based on which marker data frames will be joined.}

\item{thresh1}{Join column threshold for the first marker data frame.}

\item{thresh2}{Join column threshold for the second marker data frame.}

\item{title}{Plot title.}

\item{markersHullColor}{Color of the hull determined by \code{markers}.
Ignored if \code{markers} is \code{NULL}.}

\item{isNeg1}{Whether the first marker set represents negative markers.}

\item{isNeg2}{Whether the second marker set represents negative markers.}

\item{nameSuffix1}{Suffix appended to \code{name1} on axes and legend.}

\item{nameSuffix2}{Suffix appended to \code{name2} on axes and legend.}

\item{...}{Additional arguments passed to \code{henna::hullPlot}.}
}
\value{
A list of marker data frames.
}
\description{
This function finds upregulated or downregulate markers for Seurat object
for a given identity class and performs an additional Bonferroni correction
for multiple testing.
}
