% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_universes.R
\name{vChoose}
\alias{vChoose}
\title{Compute the prime factor decomposition of the binomial coefficient}
\usage{
vChoose(n, k)
}
\arguments{
\item{n}{Total number of elements.}

\item{k}{Number of selected elements.}
}
\value{
A vector in which positions represent prime numbers and values
represent their exponents in the factorial decomposition.
}
\description{
This function computes the prime factor decomposition of the
binomial coefficient.
}
\examples{
choose(8, 4)

}
